# api/openapi.yaml
openapi: 3.1.0
info:
  title: QianuHub 闲鱼助手 API
  version: 1.0.0
  description: 版本化 HTTP API 的唯一契约源。
servers:
  - url: /
paths:
...
  "/api/v1/cards":
    get:
      operationId: getApiV1Cards
      summary: getApiV1Cards
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: postApiV1Cards
      summary: postApiV1Cards
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardMutationRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationIDResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/cards/test-api":
    post:
      operationId: postApiV1CardsTestApi
      summary: postApiV1CardsTestApi
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardAPITestRequest'
      responses:
        '200':
          description: 测试完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardAPITestResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/cards/{card_id}":
    delete:
      operationId: deleteApiV1CardsBycard_id
      summary: deleteApiV1CardsBycard_id
      parameters:
          - name: card_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 卡密组仍被自动发货规则引用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: getApiV1CardsBycard_id
      summary: getApiV1CardsBycard_id
      parameters:
          - name: card_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: putApiV1CardsBycard_id
      summary: putApiV1CardsBycard_id
      parameters:
          - name: card_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardMutationRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/cards/{card_id}/append-data":
    post:
      operationId: postApiV1CardsBycard_idAppendData
      summary: postApiV1CardsBycard_idAppendData
      parameters:
          - name: card_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardAppendRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardAppendResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/cards/{card_id}/details":
    get:
      operationId: getApiV1CardsBycard_idDetails
      summary: getApiV1CardsBycard_idDetails
      parameters:
          - name: card_id
            in: path
            required: true
            schema:
              type: string
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/cards/batch":
    post:
      operationId: postApiV1CardsBatch
      summary: postApiV1CardsBatch
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardBatchResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
      description: 统一响应的稳定字段；业务成功字段由各阶段 operation schema 收紧。
...
    CardMutationRequest:
      type: object
      properties:
        name: {type: string}
        type: {type: string, enum: [api, text, data, image]}
        api_config:
          oneOf:
            - $ref: '#/components/schemas/APICardConfigMutation'
            - type: string
        text_content: {type: string}
        data_content: {type: string}
        image_url: {type: string}
        description: {type: string}
        enabled: {type: boolean}
        delay_seconds: {type: integer}
    CardAppendRequest:
      type: object
      required: [content]
      properties:
        content: {type: string}
    CardBatchResultResponse:
      type: object
      required: [row_no, success, name]
      properties:
        row_no: {type: integer}
        success: {type: boolean}
        id: {type: integer}
        name: {type: string}
        type: {type: string}
        error: {type: string}
    CardBatchResponse:
      type: object
      required: [success, total, created, failed, rows]
      properties:
        success: {type: boolean}
        total: {type: integer}
        created: {type: integer}
        failed: {type: integer}
        rows:
          type: array
          items: {$ref: '#/components/schemas/CardBatchResultResponse'}
    CardAppendResponse:
      type: object
      required: [success, added]
      properties:
        success: {type: boolean}
        added: {type: integer}
    MutationIDResponse:
      type: object
      required: [success, id]
      properties:
        success: {type: boolean}
        id: {type: integer}
    CardResponse:
      type: object
      required: [id, name, type, text_content, data_content, image_url, description, enabled, delay_seconds]
      properties:
        id: {type: integer}
        name: {type: string}
        type: {type: string, enum: [api, text, data, image]}
        api_config:
          allOf:
            - $ref: '#/components/schemas/APICardConfigResponse'
          nullable: true
        text_content: {type: string}
        data_content: {type: string}
        image_url: {type: string}
        description: {type: string}
        enabled: {type: boolean}
        delay_seconds: {type: integer}
        user_id: {type: integer}
    APICardConfigMutation:
      type: object
      required: [url]
      properties:
        url: {type: string, format: uri}
        method: {type: string, enum: [GET, POST], default: GET}
        timeout_seconds: {type: integer, minimum: 1, maximum: 60, default: 10}
        headers:
          type: object
          additionalProperties: true
        params:
          type: object
          additionalProperties: true
        body:
          type: object
          additionalProperties: true
        content_type: {type: string, default: application/json}
        headers_action: {type: string, enum: [retain, replace, clear]}
        params_action: {type: string, enum: [retain, replace, clear]}
        response_path: {type: string}
        retry_enabled: {type: boolean}
    APICardConfigResponse:
      type: object
      required: [url, method, timeout_seconds, content_type, retry_enabled, headers_configured, params_configured, ready]
      properties:
        url: {type: string}
        method: {type: string, enum: [GET, POST]}
        timeout_seconds: {type: integer}
        content_type: {type: string}
        response_path: {type: string}
        retry_enabled: {type: boolean}
        headers_configured: {type: boolean}
        params_configured: {type: boolean}
        ready: {type: boolean}
        validation_error: {type: string}
    CardAPITestRequest:
      type: object
      required: [api_config]
      properties:
        api_config:
          oneOf:
            - $ref: '#/components/schemas/APICardConfigMutation'
            - type: string
    CardAPITestResponse:
      type: object
      required: [status, status_code, response_content_type, response_fields]
      properties:
        status: {type: string, enum: [success, failed]}
        status_code: {type: integer}
        response_content_type: {type: string}
        response_fields: {type: array, items: {type: string}}
        extracted_value: {type: string}
        response_preview: {type: string}
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
      additionalProperties: true